Refactored extension method unit tests
