+++ Stormpath.SDK/Stormpath.SDK.Tests/Impl/Extensions/String_FromBase64_tests.cs
@@ -1,0 +1,43 @@
+﻿// <copyright file="String_FromBase64.cs" company="Stormpath, Inc.">
+// Copyright (c) 2015 Stormpath, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+// </copyright>
+
+using Shouldly;
+using Stormpath.SDK.Impl.Extensions;
+using Xunit;
+
+namespace Stormpath.SDK.Tests.Impl.Extensions
+{
+    public class String_FromBase64_tests
+    {
+        [Fact]
+        public void Returns_null_when_string_is_null()
+        {
+            ((string)null).FromBase64(System.Text.Encoding.UTF8).ShouldBeNull();
+        }
+
+        [Fact]
+        public void Returns_empty_when_string_is_empty()
+        {
+            string.Empty.FromBase64(System.Text.Encoding.UTF8).ShouldBe(string.Empty);
+        }
+
+        [Fact]
+        public void Returns_Zm9vYmFy_for_foobar()
+        {
+            "Zm9vYmFy".FromBase64(System.Text.Encoding.UTF8).ShouldBe("foobar");
+        }
+    }
+}
